ExtraDeliveryAdded

+++ Assets/Code/ProgressionModule/StorageMaterialsConfig.cs
@@ -9,10 +9,10 @@ namespace Progression
     {
         [SerializeField] private List<MaterialConfig> _storableMaterials;
         [SerializeField] private List<int> _storageCapacity;
-        [SerializeField] private List<int> _materialAddingDeltaTime;
+        [SerializeField] private List<float> _materialAddingDeltaTime;
         
         public List<MaterialConfig> StorableMaterials { get => _storableMaterials; set => _storableMaterials = value; }
         public List<int> StorageCapacity { get => _storageCapacity; set => _storageCapacity = value; }
-        public List<int> MaterialAddingDeltaTime { get => _materialAddingDeltaTime; set => _materialAddingDeltaTime = value; }
+        public List<float> MaterialAddingDeltaTime { get => _materialAddingDeltaTime; set => _materialAddingDeltaTime = value; }
     }
 }
